+++ tests/selenium_utils.py
@@ -94,6 +94,12 @@ def submit_form(d,id):
     elt=d.find_element("id", id);
     elt.submit()
 
+def click_elt_form(d,txt,sec):
+    btns=d.find_elements("class name",'ui-button')
+    btn=[b for b in btns if b.find_element("class name",'ui-button-text').text==txt][0]
+    btn.click()
+    sleep(sec)
+
 
 def select_by_text(d,id,text,sec):
     select=Select(d.find_element('id',id))